+++ boofuzz/sessions.py
@@ -49,17 +49,23 @@ class Target(object):
     Example:
         tcp_target = Target(SocketConnection(host='127.0.0.1', port=17971))
 
-    Args:
-        connection (itarget_connection.ITargetConnection): Connection to system under test.
+    :param connection: Connection to system under test.
+    :type connection: itarget_connection.ITargetConnection
+    :param repeater: Repeater to use for sending. Default None.
+    :type repeater: repeater.Repeater
+
     """
 
-    def __init__(self, connection, procmon=None, procmon_options=None, netmon=None, max_recv_bytes=10000):
+    def __init__(
+        self, connection, procmon=None, procmon_options=None, netmon=None, max_recv_bytes=10000, repeater=None
+    ):
         self._fuzz_data_logger = None
 
         self._target_connection = connection
         self.procmon = procmon
         self.netmon = netmon
         self.max_recv_bytes = max_recv_bytes
+        self.repeater = repeater
 
         # set these manually once target is instantiated.
         self.vmcontrol = None
@@ -150,10 +156,21 @@ class Target(object):
         Returns:
             None
         """
+        num_sent = 0
         if self._fuzz_data_logger is not None:
-            self._fuzz_data_logger.log_info("Sending {0} bytes...".format(len(data)))
+            repeat = ""
+            if self.repeater is not None:
+                repeat = ", " + self.repeater.log_message()
 
-        num_sent = self._target_connection.send(data=data)
+            self._fuzz_data_logger.log_info("Sending {0} bytes{1}...".format(len(data), repeat))
+
+        if self.repeater is not None:
+            self.repeater.start()
+            while self.repeater.repeat():
+                num_sent = self._target_connection.send(data=data)
+            self.repeater.reset()
+        else:
+            num_sent = self._target_connection.send(data=data)
 
         if self._fuzz_data_logger is not None:
             self._fuzz_data_logger.log_send(data[:num_sent])